Incorporate session into existing routes

+++ src/repos/UserRepo.ts
@@ -1,5 +1,6 @@
 import { getConnectionManager, Repository } from 'typeorm';
 import User from '../entities/User';
+import UserSessionRepo from './UserSessionRepo';
 
 const db = (): Repository<User> =>
   getConnectionManager()
@@ -12,9 +13,6 @@ const createUser = async (
   lastName: string,
   netID: string
 ): Promise<User> => {
-  if (!(firstName && googleID && lastName && netID)) {
-    throw Error('Invalid request body');
-  }
   const possibleUser = await db().findOne({ netID });
   if (possibleUser) {
     throw Error('User with that netID already exists');
@@ -29,27 +27,43 @@ const createUser = async (
   return user;
 };
 
-const deleteUser = async (netID: string): Promise<boolean> => {
-  const user = await getUserByNetID(netID);
-  await db().delete(user);
-  return true;
+const deleteUser = async (accessToken: string): Promise<boolean> => {
+  if (await UserSessionRepo.verifySession(accessToken)) {
+    const user = await UserSessionRepo.getUserFromToken(accessToken);
+    if (user) {
+      await db().delete(user);
+      return true;
+    } else throw Error('Could not find user with given access token');
+  } else throw Error('Could not verify session');
 };
 
 const updateUser = async (
-  currentNetID: string,
+  accessToken: string,
   firstName: string,
   lastName: string,
   netID: string
 ): Promise<boolean> => {
-  const user = await getUserByNetID(currentNetID);
-  user.firstName = firstName ? firstName : user.firstName;
-  user.lastName = lastName ? lastName : user.lastName;
-  user.netID = netID ? netID : user.netID;
-  await db().save(user);
-  return true;
+  if (await UserSessionRepo.verifySession(accessToken)) {
+    const user = await UserSessionRepo.getUserFromToken(accessToken);
+    if (user) {
+      user.firstName = firstName ? firstName : user.firstName;
+      user.lastName = lastName ? lastName : user.lastName;
+      user.netID = netID ? netID : user.netID;
+      await db().save(user);
+      return true;
+    } else throw Error('Could not find user with given access token');
+  } else throw Error('Could not verify session');
 };
 
-const getUserByNetID = async (netID: string): Promise<User> => {
+const findUser = async (accessToken: string, netID: string): Promise<User> => {
+  if (await UserSessionRepo.verifySession(accessToken)) {
+    const user = await getUserByNetID(netID);
+    if (user) return user;
+    else throw Error('User with given netID not found');
+  } else throw Error('Could not verify session');
+};
+
+const getUserByNetID = async (netID: string): Promise<User | undefined> => {
   const user = await db().findOne({
     where: { netID },
     relations: [
@@ -59,16 +73,13 @@ const getUserByNetID = async (netID: string): Promise<User> => {
       'matches.schedule.times',
     ],
   });
-  if (!user) {
-    throw Error('User with that netID does not exist');
-  } else {
-    return user;
-  }
+  return user;
 };
 
 export default {
   createUser,
   deleteUser,
+  findUser,
   getUserByNetID,
   updateUser,
 };
